Use GetUInt32Property for all properties in new RED2 that were uint32 in legacy ResourceEditInfo_t, lets see if this fails on any file

+++ ValveResourceFormat/Resource/Blocks/ResourceEditInfoStructs/ArgumentDependency.cs
@@ -24,8 +24,8 @@ namespace ValveResourceFormat.Blocks.ResourceEditInfoStructs
         {
             ParameterName = data.GetProperty<string>("m_ParameterName");
             ParameterType = data.GetProperty<string>("m_ParameterType");
-            Fingerprint = (uint)data.GetIntegerProperty("m_nFingerprint");
-            FingerprintDefault = (uint)data.GetIntegerProperty("m_nFingerprintDefault");
+            Fingerprint = data.GetUInt32Property("m_nFingerprint");
+            FingerprintDefault = data.GetUInt32Property("m_nFingerprintDefault");
         }
     }
 }
